// src/pages/Crew/types/crew.interfaces.ts
export interface FormStructureInterface {
  fullName: string;
  position: string;
  email: string;
  countryId: string;
  phone: string;
  unitIds: string[];
  order: string;
  visibleOnCall: boolean;
  visibleOnHeader: boolean;
  onCall: boolean;
  dailyReportSignature: boolean;
  emergencyContact: boolean;
  department: string;
}